+++ designPattern/src/pattern/decorator/CondimentDecorator.java
@@ -1,4 +1,4 @@
-package DecoratorPattern;
+package pattern.decorator;
 
 public abstract class CondimentDecorator extends Beverage {
     public abstract String getDescription();
