+++ designPattern/src/pattern/decorator/StarbuzzCoffee.java
@@ -1,4 +1,4 @@
-package DecoratorPattern;
+package pattern.decorator;
 
 public class StarbuzzCoffee {
     public static void main(String[] args) {
